Support arrays of BigInt, Range, Result, and Unit in entry point options (#175)
* Fix array arguments with custom handlers

* Use single-argument string constructor converters

* Seal classes

* Add tests

* Go back to using ParseArgument<T>

* Rename Result<T> to Validation<T>

* Add tests for Pauli arrays

* Add copyright notices

* Add test for repeating array option name
