+++ src/Simulation/CsharpGeneration/EntryPoint.fs
@@ -1,4 +1,7 @@
-﻿module Microsoft.Quantum.QsCompiler.CsharpGeneration.EntryPoint
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+module Microsoft.Quantum.QsCompiler.CsharpGeneration.EntryPoint
 
 open Microsoft.CodeAnalysis.CSharp
 open Microsoft.CodeAnalysis.CSharp.Syntax
@@ -45,36 +48,52 @@ let private constantsClass =
              constant "ResourcesEstimator" "string" (``literal`` AssemblyConstants.ResourcesEstimator)]
         ``}``
 
+/// The name of the C# type used by the parameter in its command-line option, given its Q# type.
+let rec private csharpParameterTypeName context (qsType : ResolvedType) =
+    match qsType.Resolution with
+    | ArrayType itemType -> sprintf "System.Collections.Generic.IEnumerable<%s>"
+                                    (csharpParameterTypeName context itemType)
+    | _ -> SimulationCode.roslynTypeName context qsType
+
 /// A sequence of all of the named parameters in the argument tuple and their respective C# and Q# types.
 let rec private parameters context doc = function
     | QsTupleItem variable ->
-        match variable.VariableName, variable.Type.Resolution with
-        | ValidName name, ArrayType itemType ->
-            // Command-line parsing libraries can't convert to IQArray. Use IEnumerable instead.
-            let typeName = sprintf "System.Collections.Generic.IEnumerable<%s>"
-                                   (SimulationCode.roslynTypeName context itemType)
+        match variable.VariableName with
+        | ValidName name ->
             Seq.singleton { Name = name.Value
                             QsharpType = variable.Type
-                            CsharpTypeName = typeName
+                            CsharpTypeName = csharpParameterTypeName context variable.Type
                             Description = ParameterDescription doc name.Value }
-        | ValidName name, _ ->
-            Seq.singleton { Name = name.Value
-                            QsharpType = variable.Type
-                            CsharpTypeName = SimulationCode.roslynTypeName context variable.Type
-                            Description = ParameterDescription doc name.Value }
-        | InvalidName, _ -> Seq.empty
+        | InvalidName -> Seq.empty
     | QsTuple items -> items |> Seq.map (parameters context doc) |> Seq.concat
 
-/// The custom argument handler for the given Q# type.
-let private argumentHandler =
-    // TODO: Support array types derived from types that need a custom argument handler.
-    function
-    | UnitType -> Some "UnitArgumentHandler"
-    | Result -> Some "ResultArgumentHandler"
-    | BigInt -> Some "BigIntArgumentHandler"
-    | Range -> Some "RangeArgumentHandler"
-    | _ -> None
-    >> Option.map (fun handler -> ``ident`` "Driver" <|.|> ``ident`` handler)
+/// The argument parser for the Q# type.
+let private argumentParser qsType =
+    let rec valueParser = function
+        | ArrayType (itemType : ResolvedType) -> valueParser itemType.Resolution
+        | BigInt -> Some "TryParseBigInteger"
+        | Range -> Some "TryParseQRange"
+        | Result -> Some "TryParseResult"
+        | UnitType -> Some "TryParseQVoid"
+        | _ -> None
+    let argParser =
+        match qsType with
+        | ArrayType _ -> "ParseArgumentsWith"
+        | _ -> "ParseArgumentWith"
+    valueParser qsType |> Option.map (fun valueParser ->
+        ``ident`` "Parsers" <.> (``ident`` argParser, [``ident`` "Parsers" <|.|> ``ident`` valueParser]))
+
+/// Adds suggestions, if any, to the option based on the Q# type.
+let private withSuggestions qsType option =
+    let rec suggestions = function
+        | ArrayType (itemType : ResolvedType) -> suggestions itemType.Resolution
+        | Result -> ["Zero"; "One"]
+        | _ -> []
+    match suggestions qsType with
+    | [] -> option
+    | suggestions ->
+        let args = option :: List.map ``literal`` suggestions
+        ``invoke`` (``ident`` "System.CommandLine.OptionExtensions.WithSuggestions") ``(`` args ``)``
 
 /// A property containing a sequence of command-line options corresponding to each parameter given.
 let private parameterOptionsProperty parameters =
@@ -86,16 +105,16 @@ let private parameterOptionsProperty parameters =
             if name.Length = 1
             then ``literal`` ("-" + name)
             else ``literal`` "--" <+> ``invoke`` toKebabCaseIdent ``(`` [``literal`` name] ``)``
-        let members =
-            argumentHandler qsType.Resolution
-            |> Option.map (fun handler -> ``ident`` "Argument" <-- handler :> ExpressionSyntax)
-            |> Option.toList
-            |> List.append [``ident`` "Required" <-- ``true``]
+        let args =
+            match argumentParser qsType.Resolution with
+            | Some parser -> [nameExpr; parser; upcast ``false``; ``literal`` desc]
+            | None -> [nameExpr; ``literal`` desc]
 
-        ``new init`` (``type`` [sprintf "%s<%s>" optionTypeName typeName]) ``(`` [nameExpr; ``literal`` desc] ``)``
+        ``new init`` (``type`` [sprintf "%s<%s>" optionTypeName typeName]) ``(`` args ``)``
             ``{``
-                members
+                [``ident`` "Required" <-- ``true``]
             ``}``
+        |> withSuggestions qsType.Resolution
 
     let options = parameters |> Seq.map getOption |> Seq.toList
     ``property-arrow_get`` optionsEnumerableTypeName "Options" [``public``; ``static``]
@@ -116,11 +135,10 @@ let private runMethod context (entryPoint : QsCallable) =
     let taskTypeName = sprintf "System.Threading.Tasks.Task<%s>" returnTypeName
     let factoryParamName = "__factory__"
 
-    let getArgExpr { Name = name; QsharpType = qsType } =
+    let argExpr { Name = name; QsharpType = qsType } =
         let property = ``ident`` "this" <|.|> ``ident`` (parameterPropertyName name)
         match qsType.Resolution with
         | ArrayType itemType ->
-            // Convert the IEnumerable property into a QArray.
             let arrayTypeName = sprintf "QArray<%s>" (SimulationCode.roslynTypeName context itemType)
             ``new`` (``type`` arrayTypeName) ``(`` [property] ``)``
         | _ -> property
@@ -128,7 +146,7 @@ let private runMethod context (entryPoint : QsCallable) =
     let callArgs : seq<ExpressionSyntax> =
         Seq.concat [
             Seq.singleton (upcast ``ident`` factoryParamName)
-            Seq.map getArgExpr (parameters context entryPoint.Documentation entryPoint.ArgumentTuple)
+            Seq.map argExpr (parameters context entryPoint.Documentation entryPoint.ArgumentTuple)
         ]
 
     ``arrow_method`` taskTypeName "Run" ``<<`` [] ``>>``
@@ -197,10 +215,12 @@ let private generatedClasses context (entryPoint : QsCallable) =
 
 /// The source code for the entry point driver.
 let private driver (entryPoint : QsCallable) =
-    let name = "Microsoft.Quantum.CsharpGeneration.Resources.EntryPointDriver.cs"
-    use stream = Assembly.GetExecutingAssembly().GetManifestResourceStream name
-    use reader = new StreamReader(stream)
-    reader.ReadToEnd().Replace("@Namespace", generatedNamespace entryPoint.FullName.Namespace.Value)
+    let source fileName =
+        let resourceName = "Microsoft.Quantum.CsharpGeneration.Resources.EntryPoint." + fileName
+        use stream = Assembly.GetExecutingAssembly().GetManifestResourceStream resourceName
+        use reader = new StreamReader(stream)
+        reader.ReadToEnd().Replace("@Namespace", generatedNamespace entryPoint.FullName.Namespace.Value)
+    String.Join (Environment.NewLine, source "Driver.cs", source "Parsers.cs", source "Validation.cs")
 
 /// Generates C# source code for a standalone executable that runs the Q# entry point.
 let generate context entryPoint =
