+++ src/Simulation/CsharpGeneration/Microsoft.Quantum.CsharpGeneration.fsproj
@@ -8,7 +8,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="Resources\EntryPointDriver.cs" />
+    <EmbeddedResource Include="Resources\EntryPoint\Driver.cs" />
+    <EmbeddedResource Include="Resources\EntryPoint\Parsers.cs" />
+    <EmbeddedResource Include="Resources\EntryPoint\Validation.cs" />
     <Compile Include="..\Common\DelaySign.fs" Link="DelaySign.fs" />
     <Compile Include="Context.fs" />
     <Compile Include="SimulationCode.fs" />
@@ -28,5 +30,5 @@
   <ItemGroup>
     <ProjectReference Include="..\RoslynWrapper\Microsoft.Quantum.RoslynWrapper.fsproj" />
   </ItemGroup>
-  
+
 </Project>
